src/core/neural_processor.py
import numpy as np

class NeuralProcessor:
    def __init__(self):
        self.model = None
        self.threshold = 0.75

    def process_data(self, data):
        # Simulate neural processing
        confidence = np.random.random()
        return {
            'processed': True,
            'confidence': confidence,
            'valid': confidence > self.threshold
        }

    def analyze_pattern(self, data):
        # Pattern analysis logic
        return {
            'pattern_detected': True,
            'risk_score': np.random.random(),
            'recommendation': 'monitor'
        } 